+++ src/components/HeadBar/HeadBar.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import MockApp from "../../fixtures/MockApp";
+import MockApp from "../../../fixtures/MockApp";
 import HeadBar from "./HeadBar";
 
 export default {
