+++ src/components/HeadBar/LocaleSwitcher/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LocaleSwitcher";
