+++ src/components/PrettyMarkdown/Heading.tsx
@@ -1,8 +1,7 @@
 import styled from "@emotion/styled";
 import * as React from "react";
 import { MOBILE } from "../../constant/mediaQuery";
-import RawText, { Props as RawTextProps, TextSize, ThemedColor, Typeface } from "../RawText";
-import RawTextThemeContext from "../RawTextThemeContext";
+import RawText, { RawTextProps, RawTextThemeContext, TextSize, ThemedColor, Typeface } from "../RawText";
 
 interface Props extends React.Attributes {
   className?: string;
