+++ src/components/PrettyMarkdown/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./PrettyMarkdown";
