+++ src/components/RawText/RawTextThemeContext.ts
@@ -1,9 +1,9 @@
 import { createContext } from "react";
-import ThemedColor from "../types/ThemedColor";
-import { TextAlignment, TextLineSize, TextSize, Typeface } from "./RawText";
-
-export { default as ThemedColor } from "../types/ThemedColor";
-export { TextAlignment, TextLineSize, TextSize, Typeface } from "./RawText";
+import ThemedColor from "../../types/ThemedColor";
+import TextAlignment from "./TextAlignment";
+import TextLineSize from "./TextLineSize";
+import TextSize from "./TextSize";
+import Typeface from "./Typeface";
 
 export interface RawTextTheme {
   color?: ThemedColor;
@@ -16,8 +16,6 @@ export interface RawTextTheme {
   italic?: boolean;
   underline?: boolean;
   lineThrough?: boolean;
-  className?: string;
-  children?: string;
 }
 
 export default createContext<RawTextTheme>({});
