+++ src/components/RawText/TextAlignment.ts
@@ -1,0 +1,8 @@
+enum TextAlignment {
+  default = "inherit",
+  start = "start",
+  end = "end",
+  center = "center"
+}
+
+export default TextAlignment;
