+++ src/components/RawText/TextLineSize.ts
@@ -1,0 +1,6 @@
+enum TextLineSize {
+  default,
+  large
+}
+
+export default TextLineSize;
