+++ src/components/RawText/TextSize.ts
@@ -1,0 +1,10 @@
+enum TextSize {
+  giantic,
+  huge,
+  large,
+  larger,
+  default,
+  small,
+}
+
+export default TextSize;
