+++ src/components/RawText/Typeface.ts
@@ -1,0 +1,7 @@
+enum Typeface {
+  body,
+  headline,
+  monospace,
+}
+
+export default Typeface;
