+++ src/components/UIText/UITextType.ts
@@ -1,0 +1,10 @@
+enum UITextType {
+  label,
+  smallLabel,
+  caption,
+  subtitle,
+  subtitle2,
+  logo,
+}
+
+export default UITextType;
