+++ src/components/UIText/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./UIText";
+export { default as UITextType } from "./UITextType";
